Refactor deepEqual helper and move to test-utils
