+++ src/specs/test-util.js
@@ -1,3 +1,5 @@
+const assert = require("assert")
+
 function errorToJson(e) {
   const props = Object.getOwnPropertyNames(e).concat('name')
   return props.reduce((p, c) => {
@@ -35,7 +37,30 @@ function sleep(ms) {
   })
 }
 
+function deepEqual(actual, expected) {
+  const a = normalizeError(actual)
+  const e = normalizeError(expected)
+  if (usingJest()) expect(a).toEqual(e)
+  else assert.deepEqual(a, e)
+}
+
+function normalizeError(v) {
+  const isError = v instanceof Error
+  if (!isError) return v
+  const props = Object.getOwnPropertyNames(v).concat("name")
+  return props.reduce((p, c) => {
+    if (c === "stack") return p
+    p[c] = v[c]
+    return p
+  }, {})
+}
+
+function usingJest() {
+  return typeof expect !== "undefined" && expect.extend && expect.anything
+}
+
 module.exports = {
   expectError,
-  sleep
+  sleep,
+  deepEqual
 }
